`QuoteArgs?`: Safer, covering more special characters.
A newly-introduced problem (of efficiency, not of correctness)
is that the resulting string is never double-quoted.
(If the original string contains single-quote characters,
it might be double-quoted and the single-quote characters would not need escaping.)
So single-quote characters are inserted in a single-quoted resulting string as `'\''`.

Happy Quote Counting !! -- as a more famous Author has said
(in the [https://duckduckgo.com/?t=ffab&q=%22The+D+Programming+Language%22+book](TDPL) Book, in 2010).

(The unittest has been updated accordingly.)